+++ src/components/icons/IconUser.vue
@@ -6,7 +6,7 @@ export default {
 
 <template>
 
-  <img src="images/User.png" height="40" width="40"/></template>
+  <img src="images/Icons/User.png" height="40" width="40"/></template>
 
 <style scoped>
