+++ src/views/LandingpageView.vue
@@ -1,5 +1,5 @@
 <script>
-import Navbar from "@/components/Navbar.vue";
+import Navbar from "@/components/NavbarHeader.vue";
 import Cards from "@/components/Cards.vue";
 
 export default {
